requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-jose[cryptography]
pwdlib[argon2]
python-multipart
pymongo
httpx
pytest
